backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
aiofiles>=23.2.0
python-multipart>=0.0.9
tiktoken>=0.6.0
python-docx>=1.1.0
